+++ src/room/participant/LocalParticipant.ts
@@ -1,5 +1,7 @@
+import { Mutex } from '@livekit/mutex';
 import {
   AddTrackRequest,
+  BackupCodecPolicy,
   ChatMessage as ChatMessageModel,
   Codec,
   DataPacket,
@@ -26,10 +28,11 @@ import {
   UserPacket,
   protoInt64,
 } from '@livekit/protocol';
+import { SignalConnectionState } from '../../api/SignalClient';
 import type { InternalRoomOptions } from '../../options';
 import { PCTransportState } from '../PCTransportManager';
 import type RTCEngine from '../RTCEngine';
-import { TextStreamWriter } from '../StreamWriter';
+import { ByteStreamWriter, TextStreamWriter } from '../StreamWriter';
 import { defaultVideoCodec } from '../defaults';
 import {
   DeviceUnsupportedError,
@@ -52,6 +55,7 @@ import LocalTrack from '../track/LocalTrack';
 import LocalTrackPublication from '../track/LocalTrackPublication';
 import LocalVideoTrack, { videoLayersFromEncodings } from '../track/LocalVideoTrack';
 import { Track } from '../track/Track';
+import { createLocalTracks } from '../track/create';
 import type {
   AudioCaptureOptions,
   BackupVideoCodec,
@@ -62,8 +66,6 @@ import type {
 } from '../track/options';
 import { ScreenSharePresets, VideoPresets, isBackupCodec } from '../track/options';
 import {
-  constraintsForOptions,
-  extractProcessorsFromOptions,
   getLogContextFromTrack,
   getTrackSourceFromProto,
   mergeDefaultOptions,
@@ -71,6 +73,7 @@ import {
   screenCaptureToDisplayMediaStreamOptions,
 } from '../track/utils';
 import {
+  type ByteStreamInfo,
   type ChatMessage,
   type DataPublishOptions,
   type SendTextOptions,
@@ -103,7 +106,6 @@ import {
   computeTrackBackupEncodings,
   computeVideoEncodings,
   getDefaultDegradationPreference,
-  mediaTrackToLocalTrack,
 } from './publishUtils';
 
 const STREAM_CHUNK_SIZE = 15_000;
@@ -527,9 +529,11 @@ export default class LocalParticipant extends Participant {
               ...this.logContext,
               ...getLogContextFromTrack(localTrack),
             });
+
             publishPromises.push(this.publishTrack(localTrack, publishOptions));
           }
           const publishedTracks = await Promise.all(publishPromises);
+
           // for screen share publications including audio, this will only return the screen share publication, not the screen share audio one
           // revisit if we want to return an array of tracks instead for v2
           [track] = publishedTracks;
@@ -611,61 +615,37 @@ export default class LocalParticipant extends Participant {
       this.roomOptions?.videoCaptureDefaults,
     );
 
-    const { audioProcessor, videoProcessor, optionsWithoutProcessor } =
-      extractProcessorsFromOptions(mergedOptionsWithProcessors);
-
-    const constraints = constraintsForOptions(optionsWithoutProcessor);
-    let stream: MediaStream | undefined;
     try {
-      stream = await navigator.mediaDevices.getUserMedia(constraints);
+      const tracks = await createLocalTracks(mergedOptionsWithProcessors, {
+        loggerName: this.roomOptions.loggerName,
+        loggerContextCb: () => this.logContext,
+      });
+      const localTracks = tracks.map((track) => {
+        if (isAudioTrack(track)) {
+          this.microphoneError = undefined;
+          track.setAudioContext(this.audioContext);
+          track.source = Track.Source.Microphone;
+          this.emit(ParticipantEvent.AudioStreamAcquired);
+        }
+        if (isVideoTrack(track)) {
+          this.cameraError = undefined;
+          track.source = Track.Source.Camera;
+        }
+        return track;
+      });
+      return localTracks;
     } catch (err) {
       if (err instanceof Error) {
-        if (constraints.audio) {
+        if (options.audio) {
           this.microphoneError = err;
         }
-        if (constraints.video) {
+        if (options.video) {
           this.cameraError = err;
         }
       }
 
       throw err;
     }
-
-    if (constraints.audio) {
-      this.microphoneError = undefined;
-      this.emit(ParticipantEvent.AudioStreamAcquired);
-    }
-    if (constraints.video) {
-      this.cameraError = undefined;
-    }
-
-    return Promise.all(
-      stream.getTracks().map(async (mediaStreamTrack) => {
-        const isAudio = mediaStreamTrack.kind === 'audio';
-        let trackConstraints: MediaTrackConstraints | undefined;
-        const conOrBool = isAudio ? constraints.audio : constraints.video;
-        if (typeof conOrBool !== 'boolean') {
-          trackConstraints = conOrBool;
-        }
-        const track = mediaTrackToLocalTrack(mediaStreamTrack, trackConstraints, {
-          loggerName: this.roomOptions.loggerName,
-          loggerContextCb: () => this.logContext,
-        });
-        if (track.kind === Track.Kind.Video) {
-          track.source = Track.Source.Camera;
-        } else if (track.kind === Track.Kind.Audio) {
-          track.source = Track.Source.Microphone;
-          track.setAudioContext(this.audioContext);
-        }
-        track.mediaStream = stream;
-        if (isAudioTrack(track) && audioProcessor) {
-          await track.setProcessor(audioProcessor);
-        } else if (isVideoTrack(track) && videoProcessor) {
-          await track.setProcessor(videoProcessor);
-        }
-        return track;
-      }),
-    );
   }
 
   /**
@@ -861,7 +841,47 @@ export default class LocalParticipant extends Participant {
     if (opts.source) {
       track.source = opts.source;
     }
-    const publishPromise = this.publish(track, opts, isStereo);
+    const publishPromise = new Promise<LocalTrackPublication>(async (resolve, reject) => {
+      try {
+        if (this.engine.client.currentState !== SignalConnectionState.CONNECTED) {
+          this.log.debug('deferring track publication until signal is connected', {
+            ...this.logContext,
+            track: getLogContextFromTrack(track),
+          });
+          const onSignalConnected = async () => {
+            try {
+              const publication = await this.publish(track, opts, isStereo);
+              resolve(publication);
+            } catch (e) {
+              reject(e);
+            }
+          };
+          setTimeout(() => {
+            this.engine.off(EngineEvent.SignalConnected, onSignalConnected);
+            reject(
+              new PublishTrackError(
+                'publishing rejected as engine not connected within timeout',
+                408,
+              ),
+            );
+          }, 15_000);
+          this.engine.once(EngineEvent.SignalConnected, onSignalConnected);
+          this.engine.on(EngineEvent.Closing, () => {
+            this.engine.off(EngineEvent.SignalConnected, onSignalConnected);
+            reject(new PublishTrackError('publishing rejected as engine closed', 499));
+          });
+        } else {
+          try {
+            const publication = await this.publish(track, opts, isStereo);
+            resolve(publication);
+          } catch (e) {
+            reject(e);
+          }
+        }
+      } catch (e) {
+        reject(e);
+      }
+    });
     this.pendingPublishPromises.set(track, publishPromise);
     try {
       const publication = await publishPromise;
@@ -963,7 +983,7 @@ export default class LocalParticipant extends Participant {
       stereo: isStereo,
       disableRed: this.isE2EEEnabled || !(opts.red ?? true),
       stream: opts?.stream,
-      backupCodecPolicy: opts?.backupCodecPolicy,
+      backupCodecPolicy: opts?.backupCodecPolicy as BackupCodecPolicy,
     });
 
     // compute encodings and layers for video
@@ -1560,6 +1580,7 @@ export default class LocalParticipant extends Participant {
       destinationIdentities: options?.destinationIdentities,
       topic: options?.topic,
       attachedStreamIds: fileIds,
+      attributes: options?.attributes,
     });
 
     await writer.write(text);
@@ -1597,6 +1618,7 @@ export default class LocalParticipant extends Participant {
       timestamp: Date.now(),
       topic: options?.topic ?? '',
       size: options?.totalSize,
+      attributes: options?.attributes,
     };
     const header = new DataStream_Header({
       streamId,
@@ -1604,6 +1626,7 @@ export default class LocalParticipant extends Participant {
       topic: info.topic,
       timestamp: numberToBigInt(info.timestamp),
       totalLength: numberToBigInt(options?.totalSize),
+      attributes: info.attributes,
       contentHeader: {
         case: 'textHeader',
         value: new DataStream_TextHeader({
@@ -1709,23 +1732,62 @@ export default class LocalParticipant extends Participant {
       onProgress?: (progress: number) => void;
     },
   ) {
-    const totalLength = file.size;
-    const header = new DataStream_Header({
-      totalLength: numberToBigInt(totalLength),
-      mimeType: options?.mimeType ?? file.type,
+    const writer = await this.streamBytes({
       streamId,
+      totalSize: file.size,
+      name: file.name,
+      mimeType: options?.mimeType ?? file.type,
       topic: options?.topic,
-      encryptionType: options?.encryptionType,
+      destinationIdentities: options?.destinationIdentities,
+    });
+    const reader = file.stream().getReader();
+    while (true) {
+      const { done, value } = await reader.read();
+      if (done) {
+        break;
+      }
+      await writer.write(value);
+    }
+    await writer.close();
+    return writer.info;
+  }
+
+  async streamBytes(options?: {
+    name?: string;
+    topic?: string;
+    attributes?: Record<string, string>;
+    destinationIdentities?: Array<string>;
+    streamId?: string;
+    mimeType?: string;
+    totalSize?: number;
+  }) {
+    const streamId = options?.streamId ?? crypto.randomUUID();
+    const destinationIdentities = options?.destinationIdentities;
+
+    const info: ByteStreamInfo = {
+      id: streamId,
+      mimeType: options?.mimeType ?? 'application/octet-stream',
+      topic: options?.topic ?? '',
+      timestamp: Date.now(),
+      attributes: options?.attributes,
+      size: options?.totalSize,
+      name: options?.name ?? 'unknown',
+    };
+
+    const header = new DataStream_Header({
+      totalLength: numberToBigInt(info.size ?? 0),
+      mimeType: info.mimeType,
+      streamId,
+      topic: info.topic,
       timestamp: numberToBigInt(Date.now()),
       contentHeader: {
         case: 'byteHeader',
         value: new DataStream_ByteHeader({
-          name: file.name,
+          name: info.name,
         }),
       },
     });
 
-    const destinationIdentities = options?.destinationIdentities;
     const packet = new DataPacket({
       destinationIdentities,
       value: {
@@ -1735,47 +1797,61 @@ export default class LocalParticipant extends Participant {
     });
 
     await this.engine.sendDataPacket(packet, DataPacket_Kind.RELIABLE);
-    function read(b: Blob): Promise<Uint8Array> {
-      return new Promise((resolve) => {
-        const fr = new FileReader();
-        fr.onload = () => {
-          resolve(new Uint8Array(fr.result as ArrayBuffer));
-        };
-        fr.readAsArrayBuffer(b);
-      });
-    }
-    const totalChunks = Math.ceil(totalLength / STREAM_CHUNK_SIZE);
-    for (let i = 0; i < totalChunks; i++) {
-      const chunkData = await read(
-        file.slice(i * STREAM_CHUNK_SIZE, Math.min((i + 1) * STREAM_CHUNK_SIZE, totalLength)),
-      );
-      await this.engine.waitForBufferStatusLow(DataPacket_Kind.RELIABLE);
-      const chunk = new DataStream_Chunk({
-        content: chunkData,
-        streamId,
-        chunkIndex: numberToBigInt(i),
-      });
-      const chunkPacket = new DataPacket({
-        destinationIdentities,
-        value: {
-          case: 'streamChunk',
-          value: chunk,
-        },
-      });
-      await this.engine.sendDataPacket(chunkPacket, DataPacket_Kind.RELIABLE);
-      options?.onProgress?.((i + 1) / totalChunks);
-    }
-    const trailer = new DataStream_Trailer({
-      streamId,
-    });
-    const trailerPacket = new DataPacket({
-      destinationIdentities,
-      value: {
-        case: 'streamTrailer',
-        value: trailer,
+
+    let chunkId = 0;
+    const writeMutex = new Mutex();
+    const engine = this.engine;
+    const log = this.log;
+
+    const writableStream = new WritableStream<Uint8Array>({
+      async write(chunk) {
+        const unlock = await writeMutex.lock();
+
+        let byteOffset = 0;
+        try {
+          while (byteOffset < chunk.byteLength) {
+            const subChunk = chunk.slice(byteOffset, byteOffset + STREAM_CHUNK_SIZE);
+            await engine.waitForBufferStatusLow(DataPacket_Kind.RELIABLE);
+            const chunkPacket = new DataPacket({
+              destinationIdentities,
+              value: {
+                case: 'streamChunk',
+                value: new DataStream_Chunk({
+                  content: subChunk,
+                  streamId,
+                  chunkIndex: numberToBigInt(chunkId),
+                }),
+              },
+            });
+            await engine.sendDataPacket(chunkPacket, DataPacket_Kind.RELIABLE);
+            chunkId += 1;
+            byteOffset += subChunk.byteLength;
+          }
+        } finally {
+          unlock();
+        }
+      },
+      async close() {
+        const trailer = new DataStream_Trailer({
+          streamId,
+        });
+        const trailerPacket = new DataPacket({
+          destinationIdentities,
+          value: {
+            case: 'streamTrailer',
+            value: trailer,
+          },
+        });
+        await engine.sendDataPacket(trailerPacket, DataPacket_Kind.RELIABLE);
+      },
+      abort(err) {
+        log.error('Sink error:', err);
       },
     });
-    await this.engine.sendDataPacket(trailerPacket, DataPacket_Kind.RELIABLE);
+
+    const byteWriter = new ByteStreamWriter(writableStream, info);
+
+    return byteWriter;
   }
 
   /**