+++ src/room/types.ts
@@ -20,6 +20,7 @@ export interface SendTextOptions {
   destinationIdentities?: Array<string>;
   attachments?: Array<File>;
   onProgress?: (progress: number) => void;
+  attributes?: Record<string, string>;
 }
 
 export interface StreamTextOptions {
@@ -31,6 +32,7 @@ export interface StreamTextOptions {
   attachedStreamIds?: Array<string>;
   replyToStreamId?: string;
   totalSize?: number;
+  attributes?: Record<string, string>;
 }
 
 export type DataPublishOptions = {